% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{genotype}
\alias{genotype}
\title{Get the genotype of a population}
\usage{
genotype(Population)
}
\arguments{
\item{Population}{A \code{\link[=Population-class]{Population}} object.}
}
\value{
A vector representing the genotype.
}
\description{
Returns the genotype stored in a \code{\link[=Population-class]{Population}} object.
}
