% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Analysis.R
\name{get_muller_plot}
\alias{get_muller_plot}
\title{Create a downloadable mutation tree plot}
\usage{
get_muller_plot(path_sim, depth, parameters, freq = FALSE, palette = NULL)
}
\arguments{
\item{path_sim}{Character string giving the path to the simulation output
directory.}

\item{depth}{Detection threshold used to filter mutations by frequency (relative to tumor size).
Only clones overcoming this threshold at least once in their lifetime will be plotted.}

\item{parameters}{A Parameters object corresponding to the parameters used in the simulation.}

\item{freq}{if TRUE, the relative abundance is use as y-axis for the plot,
if FALSE, absolute abundance is implemented.}

\item{palette}{Named character vector of colors associated with phenotype.}
}
\value{
A \code{ggplot2} object representing the mutation tree.
}
\description{
Builds a mutation tree plot from simulation output, formatted for export or
download.
}
