% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Analysis.R
\name{get_nD_index}
\alias{get_nD_index}
\alias{get_nD_index,list-method}
\title{Calculate clonal nesting (n) and clonal diversity (D) indices}
\usage{
get_nD_index(Zprovv, threshold = 10^(-2))

\S4method{get_nD_index}{list}(Zprovv, threshold = 10^(-2))
}
\arguments{
\item{Zprovv}{A list of objects of class \code{Population_with_size_nmut} representing
a snapshot of the simulated tumor.}

\item{threshold}{minimum frequency for clones to be included in the computation (relative to the whole mass)}
}
\value{
A \code{\link[tibble]{tibble}} containing two columns:
\itemize{
\item \code{D}: clonal diversity.
\item \code{n}: clonal nesting.
}
}
\description{
This function computes the Inverse Simpson Index ($D$) and the mean
mutational burden ($n$) for a simulated tumor snapshot. It filters out
clones with a relative frequency below $10^{-2}$ before calculating the metrics.
}
