% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_Params.R
\docType{class}
\name{Local_Params-class}
\alias{Local_Params-class}
\title{Local_Params class}
\description{
S4 class storing local parameters used for simulation updates at a specific
time step.
}
\section{Slots}{

\describe{
\item{\code{phenotypes_local}}{A list of local phenotypes.}

\item{\code{a_local}}{A list of local birth parameters.}

\item{\code{b_local}}{A list of local death parameters.}

\item{\code{mu_local}}{A list of local mutation parameters.}

\item{\code{rel_freq_local}}{A list of local relative frequencies.}

\item{\code{Delta}}{Numeric time increment used in the local update.}
}}

