% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population_with_size_nmut.R
\name{Ncells}
\alias{Ncells}
\title{Get the number of cells}
\usage{
Ncells(Population_with_size_nmut)
}
\arguments{
\item{Population_with_size_nmut}{A \code{\link[=Population_with_size_nmut-class]{Population_with_size_nmut}}
object.}
}
\value{
A numeric value.
}
\description{
Returns the number of cells stored in a
\code{\link[=Population_with_size_nmut-class]{Population_with_size_nmut}} object.
}
