% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{class}
\name{Parameters-class}
\alias{Parameters-class}
\title{Parameters class}
\description{
S4 class used to store simulation parameters and derived quantities for the
tumor evolution model.
}
\section{Slots}{

\describe{
\item{\code{functional_effects}}{A vector describing the functional effect associated
with each event.}

\item{\code{I}}{Number of functional events.}

\item{\code{lambda}}{Numeric vector of clone-specific proliferation advantages.}

\item{\code{mu}}{Numeric vector of clone-specific mutation rates.}

\item{\code{Competition}}{Competition matrix between clones.}

\item{\code{influence}}{Matrix describing the relative frequencies or influence of
functional events.}

\item{\code{K}}{Numeric vector of clone-specific carrying capacities.}

\item{\code{print_time}}{Numeric vector of simulation times to save or print.}

\item{\code{av_lifespan}}{Average cell lifespan.}
}}

