% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{parent}
\alias{parent}
\title{Get the parent population}
\usage{
parent(Population)
}
\arguments{
\item{Population}{A \code{\link[=Population-class]{Population}} object.}
}
\value{
A \code{\link[=Population-class]{Population}} object corresponding to the parent population.
}
\description{
Returns the parent of a population by removing the last genotype and
functional effect entry.
}
