% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{phenotype}
\alias{phenotype}
\title{Get the phenotype of a population}
\usage{
phenotype(Population)
}
\arguments{
\item{Population}{A \code{\link[=Population-class]{Population}} object.}
}
\value{
A sorted vector representing the phenotype.
}
\description{
Returns the sorted phenotype stored in a \code{\link[=Population-class]{Population}} object.
}
