% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\docType{class}
\name{Population-class}
\alias{Population-class}
\title{Population class}
\description{
S4 class representing a population, including its genotype, functional
effects, and phenotype.
}
\section{Slots}{

\describe{
\item{\code{genotype}}{Integer vector describing the genotype of the population.}

\item{\code{functional_effect}}{Integer vector describing the functional effects
associated with the genotype.}

\item{\code{phenotype}}{Integer vector describing the phenotype of the population.}
}}

