% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population_with_size_nmut.R
\name{Population}
\alias{Population}
\title{Get the population component}
\usage{
Population(Population_with_size_nmut)
}
\arguments{
\item{Population_with_size_nmut}{A \code{\link[=Population_with_size_nmut-class]{Population_with_size_nmut}}
object.}
}
\value{
A \code{\link[=Population-class]{Population}} object.
}
\description{
Returns the \code{\link[=Population-class]{Population}} object stored inside a
\code{\link[=Population_with_size_nmut-class]{Population_with_size_nmut}} object.
}
