% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Algorithm.R
\name{simulazione}
\alias{simulazione}
\alias{simulation}
\title{Run a simulation experiment.}
\usage{
simulation(
  Nexp,
  seed,
  path,
  starting_gen,
  starting_fun_eff,
  Ncells_start,
  parameters,
  tmax,
  Ncellsmax,
  epsilon_rel,
  only_last = FALSE
)
}
\arguments{
\item{Nexp}{Numeric index of the simulation experiment.
If more than one simulation is wanted, it is possible to cycle or parallelize
over this number letting it vary from 1 to a wanted Nexp_max.}

\item{seed}{Numeric seed used to initialize the random number generator
the actual seed set by the simulator is seed+Nexp to let multiple simulation
runs being replicable but avoid producing the same output repeatedly}

\item{path}{Character string giving the output directory.}

\item{starting_gen}{List of starting genotypes.}

\item{starting_fun_eff}{List of starting functional effects associated with
the starting genotypes.}

\item{Ncells_start}{Numeric vector giving the initial number of cells for
each starting population.}

\item{parameters}{A \code{\link[=Parameters-class]{Parameters}} object containing global simulation
parameters.}

\item{tmax}{Maximum simulation time.}

\item{Ncellsmax}{Optional maximum population size used as stopping criterion.}

\item{epsilon_rel}{Maximum error ammitted at each simulative step (relatively to the simulation parameters)}

\item{only_last}{logical. If TRUE only the last time stamp will we saved,
representing the ending status of the simulation.}
}
\value{
This creates a simulation output directory, saves intermediate \code{.RData} files
containing  \code{Zprovv}, which are snapshot of tumor evolution at different times and
their corresponding time (\code{time_provv}). It writes an \code{error.log} file if an error occurs.
}
\description{
Runs one stochastic simulation experiment of the population dynamics model,
starting from the initial populations and saving intermediate states to disk.
}
\details{
The simulation is restarted until a non-empty trajectory is produced. During
each run, the function repeatedly updates local parameters, samples
birth-death outcomes, applies mutation events, advances time, and stores
snapshots when checkpoint times are reached.
}
