% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{functional_effect}
\alias{functional_effect}
\title{Get the functional effects of a population}
\usage{
functional_effect(Population)
}
\arguments{
\item{Population}{A \code{\linkS4class{Population}} object.}
}
\value{
A vector representing the functional effects.
}
\description{
Returns the functional effects stored in a \code{\linkS4class{Population}} object.
}
