% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population_with_size_nmut.R
\name{Nmut}
\alias{Nmut}
\title{Get the number of mutations}
\usage{
Nmut(Population_with_size_nmut)
}
\arguments{
\item{Population_with_size_nmut}{A \code{\linkS4class{Population_with_size_nmut}}
object.}
}
\value{
A numeric value.
}
\description{
Returns the mutation count stored in a
\code{\linkS4class{Population_with_size_nmut}} object.
}
