% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population_with_size_nmut.R
\docType{class}
\name{Population_with_size_nmut-class}
\alias{Population_with_size_nmut-class}
\title{Population_with_size_nmut class}
\description{
S4 class representing a population together with its current size and the
number of mutations generated from it.
}
\section{Slots}{

\describe{
\item{\code{Population}}{A \code{\linkS4class{Population}} object.}

\item{\code{Ncells}}{Numeric value giving the number of cells in the population.}

\item{\code{Nmut}}{Numeric value giving the number of mutations generated from the
population.}
}}

